# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentación de API Saas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        }
    ],
    "paths": {
...
        "/products/{category_slug}": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Listar productos por categoría",
                "operationId": "2c26c4a58f30f7efc545832ace84f9e3",
                "parameters": [
                    {
                        "name": "category_slug",
                        "in": "path",
                        "description": "Slug de la categoría",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "zapatos"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Cantidad de productos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos de la categoría",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "category": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Obtener detalles de un producto",
                "operationId": "5d3cdbe419fcd3d89db16e83d302aa0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        },
                                        "images": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ImageProduct"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Zapatos"
                    },
                    "name_en": {
                        "type": "string",
                        "example": "Shoes"
                    },
                    "slug": {
                        "type": "string",
                        "example": "zapatos"
                    },
                    "icono": {
                        "type": "string",
                        "example": "fa-shoe-prints"
                    },
                    "imagen": {
                        "type": "string",
                        "example": "category.jpg"
                    },
                    "subCategories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "type": "string",
                        "example": "Zapato de cuero"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 150.5
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "ImageProduct": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "image": {
                        "type": "string",
                        "example": "product.jpg"
                    },
                    "description": {
                        "type": "string",
                        "example": "Vista lateral"
                    }
                },
                "type": "object"
            }
        }
    },
...
}